clean up front end code
clean up front end code
